move artifacts-dir out of daemon
